Loose fix for bug where you can go inside walls
With accelerationTimeAirborne in the GravityController script set to higher than 0.1, the player can get through the collision detection and get inside walls by hit left/right in rapid succession while next to a wall and jumping. I haven't figured out how to fix it yet with the setting greater than 0.1 (it was 0.3). I just have it set to 0.1 for now.