Possibly temporary workaround for Quit to Title from Pause menu
There is an issue at the moment such that if you use the quit button in the pause menu, the title scene loads so quickly that the user is still pressing the button, even with a relatively quick press. This input is being transferred over the scene and the result is that our default selection, the Start button, is being pressed instantly and thus restarting the game. I've implemented an artificial load time to get around this for now and I've emailed the Rewired developer to see if there is a way to mimic GetButtonDown behavior for the UI instead. If not, the artificial wait time isn't the worst.