Fixed multiple bugs with teleporting and getting into walls
Used raycasts to make sure the player will stick to the wall in any scenario, whether the projectile actually collide or if the player is really close. Also fixed a bug where the collision seemed to happen on the very last frame of the elapsed time in which the coroutine will check for a collision. This would set the collided flag but not call the collision code. This ends up not having any ill effect in terms of the collisions, but the collided flag was set to true for the next teleport, causing an instant collision. I just reset the flag at the beginning of every attack and it took care of this.